+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [WIP]
+[WIP]: https://git.smile.fr/dirtech/gdpr-dump/compare/1.0.0-beta8...master
+
+- Split functional and unit tests into two tests suites
+- Use the PSR-12 coding standard (instead of PSR-2)
+- Update the contribution guidelines
+
 ## [1.0.0-beta8] - 2019-08-19
 [1.0.0-beta8]: https://git.smile.fr/dirtech/gdpr-dump/compare/1.0.0-beta7...1.0.0-beta8